requirements.txt
PyMuPDF==1.23.8
layoutparser[paddleocr]==0.3.4
easyocr==1.7.0
transformers==4.36.2
accelerate==0.25.0
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
openai==1.7.0
tiktoken==0.5.2
Pillow==10.1.0
opencv-python==4.9.0.80
pdf2image==1.16.3
pytesseract==0.3.10
pdfplumber==0.10.3
torch==2.1.2
torchvision==0.16.2
numpy==1.24.3
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1
